move commands to pkg
