+++ pkg/cloudstorages/cloudstorage_test.go
